[Kinetics] Introduce BulkRate.h
